adicionado botão de pause na tela de AR
